Rename DeviceBase, MessageBase and StateBase to Device, Message and State
