+++ insteonplm/messages/standardSend.py
@@ -7,11 +7,11 @@ from insteonplm.constants import (MESSAGE_ACK,
                                   MESSAGE_SEND_STANDARD_MESSAGE_RECEIVED_SIZE,
                                   MESSAGE_SEND_STANDARD_MESSAGE_SIZE)
 from insteonplm.address import Address
-from insteonplm.messages import MessageBase
+from insteonplm.messages.message import Message
 from insteonplm.messages.extendedSend import ExtendedSend
 from insteonplm.messages.messageFlags import MessageFlags
 
-class StandardSend(MessageBase):
+class StandardSend(Message):
     """Insteon Standard Length Message Send 0x62."""
 
     _code = MESSAGE_SEND_STANDARD_MESSAGE_0X62
