+++ insteonplm/states/__init__.py
@@ -2,7 +2,7 @@
 import logging
 from insteonplm.address import Address
 
-class StateBase(object):
+class State(object):
     """
     Base class used by Insteon devices to hold a device state such as "Light On Level", "Temperature" or "Fan Mode".
     The class is defined with the following options:
